+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add BoxDrawingBuilder().setDefaultPad(left_spaces, right_spaces)
 - add BoxDrawingBuilder().alignLeft(columns[])
 - add BoxDrawingBuilder().alignRight(columns[])
+- add BoxDrawingBuilder().alignCenter(columns[])
 
 ## `1.1.0` 13 Jan 2023